Big changes. Users are invoiced for joining and leaving
The user invoices are cancelled if the user has not yet paid when they
leave. User creation, and deletion are fully functional and well
documented. Virtuals for the invoice provide a invoice.total property
for finding the total of an invoice. The app has a page for changing
the status of invoices so an admin or treasurer can mark who has paid
and who has not. Maybe later the invoices feature could be made into a
non-essential module for the co-op if others don't necessarily want
that feature.